% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{ests_plot}
\alias{ests_plot}
\title{Functions for plotting stochastic and stratified sampling outputs.}
\usage{
ests_plot(
  x_var,
  ests,
  g = NULL,
  plot_ci = FALSE,
  var_ests = NULL,
  title = "",
  x_name = "",
  y_name = "",
  x_labels = waiver(),
  y_labels = waiver(),
  x_lim = NULL,
  y_lim = NULL,
  ...
)
}
\arguments{
\item{x_var}{Variable to plot on the x-axis.}

\item{ests}{Variable to plot on the y-axis, e.g. stratified sampling estimates.}

\item{g}{A transformation to be applied to the \code{ests}, provided as an expression of \code{x}.}

\item{plot_ci}{Whether to plot a ribbon of the confidence intervals.}

\item{title}{Plot title.}

\item{x_name}{x-axis title.}

\item{y_name}{y-axis title.}

\item{x_labels}{x-axis labels.}

\item{y_labels}{y-axis labels.}

\item{x_lim}{x-axis labels.}

\item{y_lim}{y-axis labels.}

\item{...}{Additional inputs to be passed to \code{get_cis}.}
}
\value{
Plot of \code{ests} against \code{x_var}, with confidence intervals.
}
\description{
Functions for plotting stochastic and stratified sampling outputs.
}
