% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratsampling.R
\name{get_cis}
\alias{get_cis}
\title{Get confidence interval and (optionally) apply a transformation.}
\usage{
get_cis(estimate, var_estimate, alpha = 0.05, g = NULL, gprime = NULL)
}
\arguments{
\item{estimate}{Stochastic sampling estimate.}

\item{var_estimate}{Variance of stratified sampling estimate.}

\item{alpha}{Confidence interval level, default 0.05 to produce a 95\% confidence
interval.}

\item{g}{A transformation to be applied to the data, provided as an expression of \code{x}.}

\item{gprime}{The first derivative of g, provided as an expression of \code{x}. If not
provided, will be computed using \code{D(g, 'x')}.}
}
\value{
Upper and lower bounds of the confidence interval. If \code{g} is provided,
returns a list containing both the untransformed and transformed confidence
interval.
}
\description{
Get confidence interval and (optionally) apply a transformation.
}
