% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasim.R
\name{get_exc}
\alias{get_exc}
\title{For each threshold, get the number of exceedances in the data set.}
\usage{
get_exc(data, threshold)
}
\arguments{
\item{data}{Vector of data points of length N.}

\item{threshold}{Vector of thresholds of length M.}
}
\value{
Matrix with dimensions N*M, where entry i,j indicates whether data point i
exceeds threshold j.
}
\description{
For each threshold, get the number of exceedances in the data set.
}
