% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratsampling.R
\name{prop_alloc}
\alias{prop_alloc}
\title{Functions for stratified sampling
Proportional allocation
Allocate stratum sizes by N * w_j, where N is the total number of samples and
w_j the weight of stratum j = 1:J.}
\usage{
prop_alloc(N, weights)
}
\arguments{
\item{N}{Total number of samples.}

\item{weights}{Vector of stratum weights of length J. Must sum to one.}
}
\value{
Vector of stratum sizes.
}
\description{
Functions for stratified sampling
Proportional allocation
Allocate stratum sizes by N * w_j, where N is the total number of samples and
w_j the weight of stratum j = 1:J.
}
