% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasim.R
\name{sample_exc_data}
\alias{sample_exc_data}
\title{Sample exceedances via sampling from a Gamma distribution (to simulate an ash
concentration curve) and checking whether samples exceed a chosen threshold.}
\usage{
sample_exc_data(N, params, threshold)
}
\arguments{
\item{N}{Sample size.}

\item{params}{List containing shape and rate parameter for Gamma distribution.}

\item{threshold}{Threshold for evaluating threshold.}
}
\value{
Binary vector of length \code{N}, where 1 indicates exceedance and 0 no exceedance.
}
\description{
Sample exceedances via sampling from a Gamma distribution (to simulate an ash
concentration curve) and checking whether samples exceed a chosen threshold.
}
