% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasim.R
\name{sample_rgamma}
\alias{sample_rgamma}
\title{Functions for simulating data
Sample from a Gamma distribution.}
\usage{
sample_rgamma(size, params)
}
\description{
Functions for simulating data
Sample from a Gamma distribution.
}
