% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasim.R
\name{sample_weights}
\alias{sample_weights}
\title{Sample stratum weights randomly given a number of strata}
\usage{
sample_weights(n_strata, shape = 5)
}
\arguments{
\item{n_strata}{Number of strata.}

\item{shape}{Shape parameter of Gamma distribution with rate 1 used for generating
random weights. Defaults to 5.}
}
\description{
Sample stratum weights randomly given a number of strata
}
