% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratsampling.R
\name{strat_ests}
\alias{strat_ests}
\title{Get stratified estimates and variances given data, strata, and stratum weights.}
\usage{
strat_ests(data, strata, weights)
}
\arguments{
\item{data}{Vector of length N containing data to be aggregated.}

\item{strata}{Vector of length N indicating the stratum to which each data point
belongs.}

\item{weights}{Vector of stratum weights of length J. Must sum to one.}
}
\value{
Named list containing stratifed sampling estimate, the variance of the
estimate, and stratum means, variances, and sizes.
}
\description{
Get stratified estimates and variances given data, strata, and stratum weights.
}
