% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{stratum_means_plot}
\alias{stratum_means_plot}
\title{Plot stratum estimates and their (transformed) CIs against an x-axis.}
\usage{
stratum_means_plot(
  x_var,
  stratum_means,
  weights,
  stratum_names = NULL,
  g = NULL,
  title = "",
  x_name = "",
  y_name = "",
  legend_label = "",
  x_labels = waiver(),
  y_labels = waiver(),
  x_lim = NULL,
  y_lim = NULL
)
}
\arguments{
\item{x_var}{Variable of length N to plot on the x-axis.}

\item{stratum_means}{List of length J containing N-vectors of stratum estimates.}

\item{weights}{Vector of stratum weights of length J. Must sum to one.}

\item{stratum_names}{Vector of length J of names of strata, to construct plot legend.}

\item{g}{A transformation to be applied to the \code{ests}, provided as an expression of \code{x}.}

\item{title}{Plot title.}

\item{x_name}{x-axis title.}

\item{y_name}{y-axis title.}

\item{legend_label}{Legend title.}

\item{x_labels}{x-axis labels.}

\item{y_labels}{y-axis labels.}

\item{x_lim}{x-axis labels.}

\item{y_lim}{y-axis labels.}
}
\value{
Plot of \code{stratum_means} against \code{x_var}.
}
\description{
Plot stratum estimates and their (transformed) CIs against an x-axis.
}
