% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratsampling.R
\name{unstrat_ests}
\alias{unstrat_ests}
\title{Get unstratified sampling estimates and its variance, i.e., stochastic sampling
estimates.}
\usage{
unstrat_ests(data)
}
\arguments{
\item{data}{Vector containing data whose mean is to be estimated.}
}
\value{
Named list containing the stochastic sampling estimate and its variance.
}
\description{
Get unstratified sampling estimates and its variance, i.e., stochastic sampling
estimates.
}
